requirements.txt
streamlit
pandas
openpyxl
holidays
google-cloud-vision